+++ src/arche/rules/result.py
@@ -3,6 +3,7 @@ from enum import Enum
 from typing import Dict, List, Optional, Union
 
 import pandas as pd
+from plotly.colors import DEFAULT_PLOTLY_COLORS
 import plotly.graph_objs as go
 import plotly.io as pio
 
@@ -99,8 +100,8 @@ class Result:
 
     @property
     def figures(self):
-        if not self._figures:
-            self._figures = Result.create_figures(self.stats)
+        if not self._figures and self.stats:
+            self._figures = Result.create_figures(self.stats, self.name)
         return self._figures
 
     def add_info(self, summary, detailed=None, errors=None):
@@ -162,32 +163,25 @@ class Result:
             pio.show(f)
 
     @staticmethod
-    def create_figures(stats: List[Stat]) -> List[go.FigureWidget]:
+    def create_figures(stats: List[Stat], name: str) -> List[go.FigureWidget]:
+        if name == "Categories":
+            data = Result.build_stack_bar_data(stats)
+            layout = Result.get_layout("Category fields", len(stats))
+            layout.barmode = "stack"
+            return [go.FigureWidget(data, layout)]
         figures = []
         for stat in stats:
+            y = stat.index.values.astype(str)
             if isinstance(stat, pd.Series):
-                data = [
-                    go.Bar(
-                        x=stat.values, y=stat.index.values.astype(str), orientation="h"
-                    )
-                ]
+                data = [go.Bar(x=stat.values, y=y, orientation="h", opacity=0.7)]
             else:
                 data = [
-                    go.Bar(
-                        x=stat[c].values, y=stat.index.values, orientation="h", name=c
-                    )
+                    go.Bar(x=stat[c].values, y=y, orientation="h", opacity=0.7, name=c)
                     for c in stat.columns
                 ]
 
-            layout = go.Layout(
-                title=stat.name,
-                bargap=0.1,
-                template="seaborn",
-                height=min(len(stat) * 20, 900),
-                hovermode="y",
-                margin=dict(l=200, t=35),
-                xaxis=go.layout.XAxis(range=[0, max(stat.values.max(), 1) * 1.05]),
-            )
+            layout = Result.get_layout(stat.name, len(stat))
+            layout.xaxis = go.layout.XAxis(range=[0, max(stat.values.max(), 1) * 1.05])
             if stat.name.startswith("Coverage"):
                 layout.xaxis.tickformat = ".2p"
             if stat.name == "Coverage for boolean fields":
@@ -197,6 +191,45 @@ class Result:
 
             figures.append(go.FigureWidget(data, layout))
         return figures
+
+    @staticmethod
+    def build_stack_bar_data(values_counts: List[pd.Series]) -> List[go.Bar]:
+        """Create data for plotly stack bar chart with consistent colors between
+        bars. Each bar values have indexes unique to the bar, without any correlation
+        to other obars.
+
+        Args:
+            values_counts: an array of value_counts series
+
+        Returns:
+            A list of Bar objects.
+        """
+        data = []
+        for vc in values_counts:
+            data = data + [
+                go.Bar(
+                    x=[counts],
+                    y=[vc.name],
+                    name=str(value)[:30],
+                    orientation="h",
+                    opacity=0.6,
+                    legendgroup=vc.name,
+                    marker_color=DEFAULT_PLOTLY_COLORS[i % 10],
+                )
+                for i, (value, counts) in enumerate(vc.items())
+            ]
+        return data
+
+    @staticmethod
+    def get_layout(name: str, rows_count: int) -> go.Layout:
+        return go.Layout(
+            title=name,
+            bargap=0.1,
+            template="seaborn",
+            height=min(rows_count * 20, 900),
+            hovermode="y",
+            margin=dict(l=200, t=35),
+        )
 
     @staticmethod
     def make_annotations(stat: pd.Series) -> List[Dict]: